@@ -1,6 +1,6 @@
 {
   "name": "vue-json-pretty",
-  "version": "1.7.0-rc.2",
+  "version": "2.0.0-rc.2",
   "description": "A JSON tree view component that is easy to use and also supports data selection.",
   "author": "leezng <im.leezng@gmail.com>",
   "main": "lib/vue-json-pretty.js",
@@ -8,9 +8,9 @@
     "dev": "node build/dev-server.js",
     "build": "node build/build.js",
     "build:example": "cross-env EXAMPLE_ENV=true node build/build.js",
-    "e2e": "node test/e2e/runner.js",
-    "test": "npm run e2e",
-    "lint": "eslint --ext .js,.vue src test/e2e/specs example"
+    "build:dts": "tsc --p tsconfig.dts.json",
+    "test": "cypress open",
+    "lint": "eslint --ext .ts,.tsx,.vue src example"
   },
   "repository": {
     "type": "git",
@@ -29,49 +29,54 @@
     "url": "https://github.com/leezng/vue-json-pretty/issues"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.26.3",
-    "babel-loader": "7",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.24.1",
-    "chalk": "^2.0.1",
+    "@babel/core": "^7.12.3",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-typescript": "^7.12.1",
+    "@typescript-eslint/eslint-plugin": "^4.7.0",
+    "@typescript-eslint/parser": "^4.7.0",
+    "@vue/babel-plugin-jsx": "^1.0.0-rc.3",
+    "@vue/compiler-sfc": "^3.0.2",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "autoprefixer": "^10.0.1",
+    "babel-loader": "^8.1.0",
+    "chalk": "^4.1.0",
     "clean-webpack-plugin": "^3.0.0",
     "connect-history-api-fallback": "^1.3.0",
     "copy-webpack-plugin": "^6.0.2",
     "cross-env": "^7.0.2",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^0.28.0",
-    "cssnano": "^3.10.0",
+    "cross-spawn": "^7.0.3",
+    "css-loader": "^5.0.0",
+    "cssnano": "^4.1.10",
+    "cypress": "^5.5.0",
     "eslint": "^7.3.1",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^4.0.2",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.1.0",
+    "eslint-webpack-plugin": "^2.3.0",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.17.1",
     "file-loader": "^6.0.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "html-webpack-plugin": "^4.3.0",
-    "http-proxy-middleware": "^0.17.3",
+    "http-proxy-middleware": "^1.0.6",
     "less": "^3.11.3",
-    "less-loader": "^6.1.2",
-    "mini-css-extract-plugin": "^0.9.0",
-    "nightwatch": "^1.0.19",
-    "opn": "^5.1.0",
+    "less-loader": "^7.0.2",
+    "mini-css-extract-plugin": "^1.2.1",
+    "open": "^7.3.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "prettier": "^2.1.2",
     "selenium-server": "^3.0.1",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.6",
+    "semver": "^7.3.2",
+    "shelljs": "^0.8.4",
+    "typescript": "^4.0.5",
     "url-loader": "^4.1.0",
-    "vue": "^2.6.11",
-    "vue-loader": "^14.2.4",
+    "vue": "^3.0.2",
+    "vue-loader": "^16.0.0-rc.1",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.11",
-    "webpack": "^4.43.0",
+    "webpack": "^5.3.0",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^5.2.0"
   },
   "engines": {
     "node": ">= 10.0.0",
@@ -83,7 +88,8 @@
     "not ie <= 11"
   ],
   "files": [
-    "lib"
+    "lib",
+    "types"
   ],
   "dependencies": {}
 }
