@@ -1,8 +1,8 @@
 // https://github.com/michael-ciniawsky/postcss-load-config
 
 module.exports = {
-  "plugins": {
+  plugins: {
     // to edit target browsers: use "browserslist" field in package.json
-    "autoprefixer": {}
-  }
-}
+    autoprefixer: {},
+  },
+};
