@@ -1,6 +1,6 @@
-const merge = require('webpack-merge')
-const prodEnv = require('./prod.env')
+const { merge } = require('webpack-merge');
+const prodEnv = require('./prod.env');
 
 module.exports = merge(prodEnv, {
-  NODE_ENV: '"development"'
-})
+  NODE_ENV: '"development"',
+});
