@@ -1,6 +1,6 @@
-const merge = require('webpack-merge')
-const devEnv = require('./dev.env')
+const { merge } = require('webpack-merge');
+const devEnv = require('./dev.env');
 
 module.exports = merge(devEnv, {
-  NODE_ENV: '"testing"'
-})
+  NODE_ENV: '"testing"',
+});
